Provides the visual GUI for the program
The GameGUI will set up the while interface using the library java.awt and javax.swing.
The game will be in a small pop up window, and this file is defining the look.
Notice, there is a public method updateCell provided at the end. It is used for checking
if the chosen cell contains a ship or just water and indicates the result with a colour.